routes compet

+++ app/Http/Controllers/CompetitionController.php
@@ -131,14 +131,19 @@ class CompetitionController extends Controller
 
     public function desinscrire($evenement, $categorie, $animal)
     {
-   
+        
        $resultat = Resultat::where('animal_id',$animal)->where('evenement_id',$evenement)->where('categorie_id', $categorie);
-    
-       if ($resultat->delete()) {
+       $date = $resultat->evenement->date_create_immutable;
+       if (Gamedata::date() >= $date){
+            if ($resultat->delete()) {
        
        
         return redirect()->back()->with('status',"votre animal a été désinscrit");
     }
+    else {
+        return redirect()->back()->with('status',"Trop tard, la compétition a eu lieu");
+    }
+}
        
     }
